+++ test/auth/oidc_test.go
@@ -150,21 +150,6 @@ func TestSequenceSupportsOIDC(t *testing.T) {
 	env.Test(ctx, t, oidc.SequenceHasAudienceOfInputChannel(name, env.Namespace(), channel_impl.GVR(), channel_impl.GVK().Kind))
 }
 
-func TestApiserversourceSendEventWithJWT(t *testing.T) {
-	t.Parallel()
-
-	ctx, env := global.Environment(
-		knative.WithKnativeNamespace(system.Namespace()),
-		knative.WithLoggingConfig,
-		knative.WithTracingConfig,
-		k8s.WithEventListener,
-		environment.Managed(t),
-		eventshub.WithTLS(t),
-	)
-
-	env.Test(ctx, t, oidc.ApiserversourceSendEventWithJWT())
-}
-
 func TestContainerSourceSendsEventsWithOIDCSupport(t *testing.T) {
 	t.Parallel()
